Targeted dive ability for Shadow Team.
