+++ mods/ca/rules/infantry.yaml
@@ -3769,9 +3769,9 @@ SGLI:
 		InitialFacing: 192
 		TurnSpeed: 16
 		Speed: 115
-		RepulsionSpeed: 40
 		MaximumPitch: 56
 		Voice: Move
+		Repulsable: false
 	Armament:
 		Weapon: ShadowGliderGrenade
 		LocalOffset: 0,0,-85
@@ -3798,9 +3798,7 @@ SGLI:
 		DecorationBounds: 1280, 1194, 0, 85
 	TransformOnCondition:
 		IntoActor: sgli.deployed
-		RequiresCondition: deployed
-	GrantConditionOnDeploy:
-		DeployedCondition: deployed
+		RequiresCondition: exhausted
 	Targetable@AIRBORNE:
 		RequiresCondition: airborne && !being-warped
 		TargetTypes: AirSmall
@@ -3810,8 +3808,8 @@ SGLI:
 	TimedConditionBar@GLIDING:
 		Condition: gliding
 		Color: 8A8A8A
-	GrantCondition@AUTODEPLOY:
-		Condition: deployed
+	GrantCondition@Exhausted:
+		Condition: exhausted
 		RequiresCondition: !gliding
 	RejectsOrders:
 		RequiresCondition: !gliding
@@ -3821,6 +3819,12 @@ SGLI:
 	-Contrail@2:
 	-ActorLostNotification:
 	-SoundOnDamageTransitionCA:
+	TargetedDiveAbility:
+		MinDistance: 3
+		MaxDistance: 8
+		Speed: 175
+		TransformIntoActor: shad
+		ShowSelectionBar: false
 
 SGLI.crashing:
 	Inherits: SGLI
@@ -3843,11 +3847,10 @@ SGLI.crashing:
 	-DeathSounds:
 	-UpdatesPlayerStatistics:
 	-SpawnActorOnDeath:
-	-GrantConditionOnDeploy:
 	-TransformOnCondition:
 	-GrantTimedCondition@GLIDING:
 	-TimedConditionBar@GLIDING:
-	-GrantCondition@AUTODEPLOY:
+	-GrantCondition@Exhausted:
 	AmbientSoundCA:
 		SoundFiles: chute1.aud
 		Interval: 6000
@@ -3877,11 +3880,10 @@ SGLI.deployed:
 	AmbientSoundCA:
 		SoundFiles: chute1.aud
 		Interval: 6000
-	-GrantConditionOnDeploy:
 	-TransformOnCondition:
 	-GrantTimedCondition@GLIDING:
 	-TimedConditionBar@GLIDING:
-	-GrantCondition@AUTODEPLOY:
+	-GrantCondition@Exhausted:
 	RejectsOrders:
 		-RequiresCondition:
 	-MapEditorData:
