+++ mods/ca/weapons/ballistics.yaml
@@ -784,7 +784,7 @@ ShadowOperativeGrenade:
 
 ShadowGliderGrenade:
 	Range: 1c0
-	ReloadDelay: 80
+	ReloadDelay: 100
 	Inherits: ShadowOperativeGrenade
 	-InvalidTargets:
 	Projectile:
